teacher_assistant/pyproject.toml
[project]
name = "teacher_assistant"
version = "0.1.0"
description = "AI-powered teacher assistant using CrewAI"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
teacher_assistant = ["config/*.yaml"]
[project.scripts]
run-crew = "teacher_assistant.main:run"